requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
openai>=1.60.0
gitpython>=3.1.40
pyyaml>=6.0.1
sqlalchemy>=2.0.0
aiofiles>=24.1.0
httpx>=0.27.0
pillow>=10.0.0
python-multipart>=0.0.9
aiosqlite>=0.20.0
greenlet>=3.0.0
